+++ src/main/java/team/gwon/haveameal/member/controller/MemberController.java
@@ -1,12 +1,12 @@
-package team.gwon.haveameal.register.controller;
+package team.gwon.haveameal.member.controller;
 
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
-import team.gwon.haveameal.register.domain.MemberRegister;
-import team.gwon.haveameal.register.service.MemberService;
+import team.gwon.haveameal.member.domain.MemberRegister;
+import team.gwon.haveameal.member.service.MemberService;
 
 @RestController
 @RequiredArgsConstructor
