+++ src/main/resources/mapper/MemberMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="team.gwon.haveameal.register.mapper.MemberMapper">
-    <insert id="insertMember" parameterType="team.gwon.haveameal.register.domain.MemberEntity">
+<mapper namespace="team.gwon.haveameal.member.mapper.MemberMapper">
+    <insert id="insertMember" parameterType="team.gwon.haveameal.member.domain.MemberEntity">
         INSERT INTO member
         VALUES (#{member_id}, #{password}, #{name}, #{phone}, #{role}, #{withdrawal}, #{withdrawal_at})
     </insert>
